[mega-merge-8] final integration + performance + coverage + documentation + quality final
Integration: pipeline executor v3, event bus, config validator, output manager, runtime module loader
Performance: request pool, payload cache, response dedup, smart scheduler, memory-efficient finding store
Coverage: DOM clobbering, client-side template injection, insecure file handling, CORS v3, security features
Documentation: doc comments on ALL public types across ALL crates
Quality: 61 orphaned modules wired, 0 clippy warnings, 0 errors, all tests passing

+++ crates/evasion-engine/src/tls_attacks.rs
@@ -226,8 +226,8 @@ fn build_client_hello(
     // Extensions: SNI
     let sni_bytes = sni.as_bytes();
     let sni_list_len = (sni_bytes.len() + 3) as u16;
-    let sni_ext_len = (sni_list_len + 2) as u16;
-    let extensions_len = (sni_ext_len + 4) as u16;
+    let sni_ext_len = sni_list_len + 2;
+    let extensions_len = sni_ext_len + 4;
 
     hello.extend_from_slice(&extensions_len.to_be_bytes());
     // SNI extension type = 0x0000
